krb-udp-cleanup-20010119
Make krb_udp try kerberos-iv as a fallback (from the original default of
kerberos4) before going to "kerberos". "kerberos" incidentally is 88 on all
modern platforms, so lord forbid you want to run this and something which
can actually answer krb5 requests, if you're using the default services.

Also minor fix for the size of an answer packet.