STABLE10-honor-solaris-attr-hint-20010112
"The problem that it fixes is as follows: suppose user A's process maps
something into its address space out of AFS. Now if user B runs "ps",
or otherwise reads /proc/NNN/status, the kernel code will lock the
associated entry in /proc and then proceed to find the size of the
address space for that process; in doing so, it calls getattr(). If B
does not have tokens to access the file, the fileserver will eventually
start delaying responses to FetchStatus requests, and the kernel will
block with /proc locked.

The user-visible effect is that "ps", "w" and so on start blocking
uninterruptibly for long periods of time."


(cherry picked from commit 22293eeada64329250864b4f46dfce3ca857f674)