bosserver-syslog-option-including-selectable-facility-20010315
"Simple patch to add "-syslog" option to bosserver. Causes it to direct all
logging activity to syslogd instead of to logs/BosLog"

currently unimplemented for NT

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
Don't use a static outside its scope