first-pass-at-alpha-linux-support-20010312
Changes for Alpha to previous 64 bit changes for sparc (sparc64 isn't
necessarily analogous to alpha or any other 64 bit linux platform)

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
use mkstemp

====================
Include necessary file for alpha config

====================
Deal with 64 bit time_t

====================
Don't do increments in a way that can be miscompiled

====================
configure for alpha linux

====================
deal with time and initialize a variable

====================
don't use ViceIoctl32 for alpha linux

====================
deal with compiling kdump for alpha linux

====================
mods in callback.c analogous to dux port; mods in viced to deal with
thread creation function return convention

====================
config for linux alpha

====================
ids for alpha linux consistent with new-style sysname grouping

====================
makefile and config files for new-style sysnames

====================
MakefileProto changes for alpha linux with new-style sysnames. A change
in config/mc.c will be needed for this to configure correctly, and it will
be added soon

====================
note alpha port and the state it's in to hopefully avoid questions