DEVELAC-junk-delta-20010327
push additions so i can cvs update elsewhere

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

start on genericification to "features" which at this stage are hardcoded to
platforms

====================

placeholder file with per-platform features selection info

====================

install placeholder include file for now