restore-writelogbuffer-for-mrafs-20010514
MR-AFs is making use of this to hook into logging to write multiline messages
into the log

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

move forward to current version of function