linux24-use-flock64-instead-of-flock-20010604
"locking on linux 2.4.x does not work. 2.4.x uses internally 64bit offsets
(start and end of lock) and end is set to 64 bit int max. When it converts
to struct flock, len is set to 0xffffffff (cut off high 32 bits) and afs_lockctl
decides it is byte-range locking. So on 2.4.x kernel there should be used flock64
instead of flock (flock64 is typedef to flock on 64bit systems such as sparc64
and ia64)"

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

use flock64 for linux24