afsconfig-and-rcsid-all-around-20010705
convert rest of source to afsconfig; include rcsid macros

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

remove bogus if/define/endif triple

====================

revert non-rcsid and afsconfig portion of patch