configure-guess-afs-sysnames-20010711
make configure attempt to guess the correct afs sysname for your platform

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

use precomputed kernel we're building for not uname output

====================

stick in missing systems