afs-vcache-state-flag-fix-and-cleanup-20010806
"There appears to be a conflict between two vcache state flags:
CWritingUFS and CDCLock are both 0x200000.  This patch moves
CDCLock to 0x2000000 (which I'm guessing was the intended value
from the position of that #define).  It also reformats the block
of #define's to be consistent in number format and indentation,
to make it somewhat more readable, and to make such conflicts
less likely to reappear."