dux-create-loadable-kernel-module-20010829
with minor correction by shadow@dementia.org to install the module, but
we now create a loadable kernel module for dux

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

actually build afs.mod

====================

fix typo