more-makefiles-cleanup-20010829
get rid of cases where we'd end up with two / because of DESTDIR pasting

macroize calls to rm

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

get rid of another case of bogus trailing quoting