config-makefile-further-cleanup-20010830
ordering of dependancies caused ${DESTDIR}/include/afs to not exist in
time, masked if you built into an existing destdir. so let's make
the makefile simpler and more logically paraallel.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

get missing backquote