misc-build-cleanup-20010917
1) Removal of unused variables
2) some NT ifdef cleanup
3) use of size_t rather than int to store the result of strlen()
4) some bugs that really only occured when errors occured.
5) some functions were implicitly declared as 'int' but doing a
   "return;" Some I declared as 'void' and prototyped them; some I
   just returned a value.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

put symbol in an ifndef for winnt

====================


change erroneous void to int

====================

technically the cmdprocs return int

====================


update function to be void