allow-namei-fileservers-to-use-directories-instead-of-partitions-for-vicep-20010924
 -- /vicepX/AFSIDat/README is created the first time the fileserver
    is started, rather than the second time.

 -- For the namei fileserver, /vicepa/Lock/vicepa is created as
    the lockfile.  Previously, on Solaris, the raw device was
    was being used for LockPartition(), and that breaks when
    there isn't a corresponding device.

 -- The VolPartitions and XVolPartitions volserver calls now return
    the list of attached (rather than mounted) partitions.  This is
    consistent with the NT fileserver and removes duplicate code.

The actual magic for allowing the fileserver to attach non-mounted
partitions is to create /vicepX/AlwaysAttach.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

carefully avoid non vicep directories and/or partitions as it may get upset at
automounters or running cachemanagers!

====================

more carefully avoid non-vicepX partitions