logging-enhancements-20011009
The bosserver now logs the exit status of processes under its control.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Messages generated as the result of assertion failures now include
timestamps.