linux-rc-file-cleanup-20011008
Modularize the Linux RC file(s) a bit.  In particular, add more
options for larger caches, allow autogeneration of the cacheinfo file,
make sure AFS only starts if the network is running, add SuSE header
information, and cope with the Red Hat initscripts better (on Red Hat
only, but should work on all systems).

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

fix silly typo