linux-osi-sleep-avoid-forgetting-events-20020101
based on suggestion from Ted Anderson
"the changes make more sense
than the code as it currently exists.  the only think i am nervous
about is the dontSleep delete.  while it makes more sense to just
not wakeup sleepers if none exist, i suppose its possible that
some bit of afs code wants acausal (wake before sleep) events.
that does seem quite unlikely.  just looking at the sleep on
solaris, it checks the seq number to get the next event not
a previous event.

i imported the changes and make the fixup in osi_stoplistener().
i dropped some of the silly syntax changes that junked up the
diff -- this makes it a bit easier to see what was changed.
i just added an assert in afs_addevent for quality assurance
purposes."

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

fix for osi_StopListener so it does the rigth thing