solaris-obtain-locks-under-GLOCK-20020101
Fix the two places in Solaris afs_GetOnePage where locks
  are obtained/released without GLOCK.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Fix the two places in Solaris afs_GetOnePage where locks
  are obtained/released without GLOCK.

(Fix error in application of patch)