rx-cleanup-gettimeofday-use-20020120
don't blindly make assumptions about gettimeofday

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

redo in a way that doesn't screw up with respect to libafsrpc and libafsrpc-
using suites

====================

redo in a way that doesn't screw up with respect to libafsrpc and libafsrpc-
using suites