source-map-20020121
with modifications by shadow@dementia.org

a guide to what's where in the OpenAFS source

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

with modifications by shadow@dementia.org

a guide to what's where in the OpenAFS source

this time the right file. a sure sign that it's bedtime.