sys-add-aix-syscall-exports-removed-incorrectly-earlier-20020319
to export syscalls as functions you need the magic exp file

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

and also reinstall them for the user cache manager library