suggested by lha@stacken.kth.se
drop locks when rxi_Alloc fails, and rework SRXAFSCB_GetCellServDB to use
afs_GetCellByIndex. He suggested same for GetLocalCell, and as long as cell 1
and not CPrimary is the local cell indicator that works, but I left that part
of the code as is for now.


(cherry picked from commit fc09b92dce32025c8e1ac762b3f8f07bc731869a)