PROTO13-remove-osi-dnlc-prototype-include-20020506
Undo need-prototype-for-osi-dnlc-lookup-20020503, as it created
double inclusion of afs_osidnlc.h.

An equivalent patch was already present in the mainline as part of
linux-dcache-inconsistency-fix-20020115, but its pullup to the stable
branch, STABLE12-linux-dcache-inconsistency-fix-20020120, was oddly
missing this particular part of the diff.


(cherry picked from commit 6aae6c02849dfed30fab3a407ff445353f1c5e31)