prototypes-fixes-20020821
fix the following problems
- including osi_vfs.h on almost all platforms, even though afsincludes.h
already deals with it
- universally declaring afs_globalVFS as a struct vfs *
- declaring afs_stats_XferSumBytes in a header without a storage class
- using afsincludes.h without sysincludes.h
- make clean removes a source file in rxkad

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

irix needs this, put it back

====================

cast parm argument to CreateProcess

====================

more lwp createprocess fun

====================

more lwp