salvager-dont-mask-uniq-comparison-in-vnodes-only-in-inodes-20020920
the masking operation is for inodes when using "inode spares" fields which
overflow, not for vnode operations.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================


remove an extraneous right paren