solaris-nfs-translator-expand-macros-and-cleanup-20021011
trying to get this to work

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

make it not go down in flames

====================

fix nfsv2 also

====================

looks like we need to do this when we create something new also

====================

cast correctly.

====================

and add in some fstrace crap which afs_trace.et already had