rxkad-krb5-20021027
Added support to rxkad for servers accepting Kerberos V5 tickets and
truncated "proposal 2b" tickets.  When used with an appropriate aklog
or krb524d (such as shipped with Heimdal and MIT Kerberos), this allows
the use of Kerberos V5 tickets (with DES session keys) to authenticate
connections to OpenAFS servers.