fssync-interface-change-to-deal-with-callback-delays-20021204
move callback breaking into a separate thread

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

the rest of it. somehow the original version of this was lost.

====================

the rest of it. somehow the original version of this was lost.

====================

the rest of it. somehow the original version of this was lost.