callbacks-later-linked-list-fix-20030114
the linked list code was never updated for multiple passes when we
started doing that.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================


the linked list code was never updated for multiple passes when we
started doing that.