amd64-hook-ia32-table-20030521
keep track of the ia32 table separately. annoying but needed.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

provide the variables we need for ia32 sys call table if
sys_call_table is exported