ptprocs-supergroups-stub-20030811
have a stub to go with the unifdef'd .xg file.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

have a stub to go with the unifdef'd .xg file.

make the stub actually work. i wonder if it's ok to cheat like this.
(return RXGEN_OPCODE)