openbsd-20031007
use poll instead of select for 3.5 and beyond
reported by provos

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

works much better if you check the right vnode pointer