cachemgr-rx-multi-gettime-20031109
use multi Rx to GetTime all servers we wish to talk to in parallel. This
should have the effect of compressing the amount of time to time out all
fileservers to (1 times full set of Rx transmit intervals) instead of (n times
same). VLservers get handled out of band and so aren't in scope for this
change.