namei-ihandle-leak-20031114
FIXES 2518

namei_dec now logs if you have open ihandles when you unlink a file

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 2518

flush all dir buffers we have cached when rewriting a volume from a restore

avoids holding an ihandle which has an open fdhandle for an unlinked file

====================
FIXES 2518

remove logging code for now. it appears to trigger on some possibly legitimate operations.