volser-restore-create-timestamp-not-touched-on-incremental-20031114
FIXES 2521

        The issue: when you restore an incremental dump to an existing volume,
the creation timestamp of the target volume gets reset to the current timestamp
(literally the timestamp of when the 'vos restore' command is being executed).
That messes things up for any subsequent 'vos release' operation, because such
operation then needs to do a full release rather than an incremental one.  For
large volumes that can be quite an issue.