thread-and-timer-corrections-to-afscreds-20031121
Changes to afscreds to place the obtain tokens dialog into a separate
thread to prevent blocking of the Windows Message queue.  Requires
utilization of mutex semaphores to protect credential data structures.

Previous versions of afscreds would set/unset the timer event every
time the program received an event indicating user activity including
dragging the mouse across the systemtray icon.  This resulted in
extremely unpredictable behavior.  Now the timer event is only turned
off when it must be turned off; and turned on when it must be turned on.
The result is a credential expiration check once a minute.