nsis-patches-20031211
To do list:

* remove /wizard from afssvrcfg.exe shortcuts

* add a "runonce" for afssvrcfg.exe /wizard when the Server is
installed

* enforce a restriction of no Freelance mode when the Server
is installed

* enforce a requirement of the Client install when the Server
is installed

* add a control panel shortcut to afssvrcfg.exe

* remove "recommended" tag from the AFS Server option

>>>>>>>>>>>>>>>>>>>>>

This is everything we talked about, plus:

1) Fixes the installer so re-installation behaves as expected.
All existing settings are preserved. (With the exception of what
is modifiable during install.  The cell name is prepopulated, but
the user CAN change it)
2) Likewise, an upgrade will retain previous settings.  (There
might be a point from which you cannot upgrade, i.e., from IBM
3.4 or something)
3) Variables named consistently, more code left out if NSIS 2.0b4.

-Rob