nsis-do-not-install-sdk-with-client-20040101
From Rob Murawski: This is a diff with the changes we talked about for
making a separate section for the SDK files.  It is not installed by
default, unless the SDK files already exist.  (To keep them consistent
on the machine)

Also, a fix to the ShouldClientInstall function to properly handle
downgrades.