smb_fid_t-leak-20040229
There was a significant leak of smb_fid_t objects produced when copying
files from an AFS volume.  This is clearly part of the problem reported
by UNCC.  However, there is still an unexplained memory leak associated
with reading and writing files even if the data is only being read from
the cache.

I am noticing reports of memory allocation overruns from the Debug CRT
when freeing SCache objects.  This could be a side effect of whatever is
causing the memory leak.  This will require more research.