service-dependencies-20040315
Fix the location of the registy values to be used for specifying
the service dependencies of the AFS Network Provider.  AFS is
dependent on PNP_TDI (tcpip, netbt, etc); RpcSs; and NetBios.

Upgrades were failing to read the Cell name from the proper location
so that it could be restored.