registry-20040316
Fix the DependsOnGroup NetworkProvider key to ensure it is placed in
the correct location

Change the location of the ShowTrayIcon key so that we no longer store
user data in the HKLM Transarc tree

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Move the locations of the ShowTrayIcon and cell name based expiration
reminder keys out of the HKLM\SOFTWARE\Transarc... key and into the
{HKCU,HKLM}\SOFTWARE\OpenAFS\... tree.  This is safe to do because these
keys are transient in nature (created by afscreds.exe) and really do
belong to the user and not to the machine.