reg-expand-str-20040316
Change the NetbiosName registry value from REG_SZ to REG_EXPAND_SZ
and add the necessary code to expand the strings.   This will allow
the use of %COMPUTERNAME%-AFS in case people want to explicitly use
a non-portable name.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Update text for NetbiosName value.