move-readme-to-top-20040317
Move the README files to the top directory where they belong.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

As of 1.3.60 move the Windows Readme files from src to top directory