irix-attempt-to-detect-kernel-mem-routines-20040317
FIXES 3651

first pass at routine to detect kernel mem* routines

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 3651

second try

====================
FIXES 3651

third try

====================
FIXES 3651

this time for sure.
includes the definitions too, this way we can not have macros conflict with
headers we don't need

====================
FIXES 3651

this time for sure.
includes the definitions too, this way we can not have macros conflict with
headers we don't need

====================
FIXES 3651

this time for sure.
includes the definitions too, this way we can not have macros conflict with
headers we don't need