VLCallBack-vos-20040321
The variable 'm_unique' is an afs_uint32 where the desired parameter
must be of type struct VLCallBack when attempting to call VL_GetAddrs.
I do not know how this code ever worked on any platform as the stack
was being overwritten.

This produced binary, vos.exe on windows, also generates massive
memory leaks.  I do not think we care all that much because the
utility does not stay around all that long.  However, the fact that
the API makes it so easy to lose memory is scary.