nsis-installer-20040321
Remove Rob's new way of installing services as it does not appear
to work properly.

Add in code to test the legitimacy of the upgrade to ensure we don't
try to upgrade an old IBM AFS version that isn't supported

Make sure that the only AFS in the PATH is ours and that we are only
in once.

Add the TransarcAFSDaemon entry into the Control\NetworkProviders::Order value
to support Integrated Login

Build service.exe and killer.exe with static libraries to avoid
run-time dependencies at install and uninstall time

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Check both Client and Server for unsupported previous installation