windows-srvr-cfg-resource-20040401
Renumber and re-organize all of the String Table entries
to adhere to the rules specified for dynamically loading
them with the tools in WINNT/talocale.  All numbers must
be sequential.  Each block must begin on a mod 16 == 0
value.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Update resource constants to avoid conflicts