fixes-from-andrei-20040427
Cleanup the makefile to remove unnecessary defaultlib restrictions

Remove the DST fix added to address Rodney's bug with the timezone being
off by an hour on files one hour after the time change.  Still need to
identify the real cause.

Finally fix once and for all the afsd_service.exe shutdown error.
The ServiceHandler must be declared with the WINAPI calling convention
because it returns a value.