callback-rxcon-move-20040602
FIXES 4891

apply it disabled in the fileserver, but provide mech for client to suggest a particular IP address to the fileserver

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 4891

apply it disabled in the fileserver, but provide mech for client to suggest a pa
rticular IP address to the fileserver

====================
FIXES 4891

apply it disabled in the fileserver, but provide mech for client to suggest a pa
rticular IP address to the fileserver

====================
FIXES 4891

apply it disabled in the fileserver, but provide mech for client to suggest a pa
rticular IP address to the fileserver