darwin-updates-20040613
complain if TryReclaim gets a vnode which is on the inactive list

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

be willing to instantiate a ubc in our lookup vop (rename needs this so
the followup vget doesn't get a vnode with no ubc)

refuse to instantiate a ubc if VORECLAIM is set

rework rename 2 links to same vnode case. (yes, you can have hard links in AFS)

====================

drop the parent directory ref and remove name from the kernel name cache
when recycling a vnode.