fs-sysname-20040619
Begin the process of supporting multiple sysnames on Windows as is
done on Unix.  These changes replace the internal variable with a list
of sysnames.

Still to do:  change command line interface to allow entry of multiple
names; and replace cm_ExpandSysName() with code similar to Unix which
allows expansion of one element in the list at a time.