new-loopback-dll-20040622
Construct a new afsloopback.dll which contains the routines
for installing, removing, and verifying the existance of
a loopback adapter.  This dll will be used by both the NSIS
and the Wix installers.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

Update the wix installer to use the new version of instloop.exe
which uses the new afsloopback.dll