loopback-20040623
properly set dependencies for NSIS and wix targets to build loopback target

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

do not install a second loopback adapter if one is already installed.

====================

1.  Custom actions should not depend on any library that is not in the Windows distribution.  Change makefile to link
with a static runtime.

2.  Add common reporting mechanism to report ActionData messages back to the MSI process during the loopback
installation.

3.  CoInitializeSecurity can be called only once per process.  When running as a custom action DLL under the MSI process
we won't be able to successfully call this since the MSI process beats us to it.

====================



Compensate for difference in argument passing in MSI and RunDll32