submounts-20040720
Fix afsshare.c to strip the MountRoot off the afs-path before inserting into
the registry

Fix smb.c to read the submounts out of the registry.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

fix submount extraction from registry