misc-patches-20040726
Update .cvsignore files for windows

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

cleanup uninitialized variables

====================

more dlls to be replaced

====================

When determining the full path of a UNC path, strip the server and
share names