rx-provide-binding-version-of-init-20040728
i should be dragged away and beaten for this

first pass at interface to allow bound Rx sockets

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

i should be dragged away and beaten for this

====================

i should be dragged away and beaten for this

first pass at interface to allow bound Rx sockets

====================

i should be dragged away and beaten for this

first pass at interface to allow bound Rx sockets

====================

i should be dragged away and beaten for this

first pass at interface to allow bound Rx sockets

====================

i should be dragged away and beaten for this

first pass at interface to allow bound Rx sockets

====================

i should be dragged away and beaten for this

first pass at interface to allow bound Rx sockets

====================

i should be dragged away and beaten for this

first pass at interface to allow bound Rx sockets