volser-restore-timestamp-cleanup-20040728
FIXES 5926

This patch solves the problem of how timestamps are handled by during a
'vos restore' by allowing the specification of there the creation and last
updated timestamps are taken from:

        - generate a new one using the current time
        - retain the one in the target volume if there is one
        - restore the timestamp from the dump file

This is needed to allow sequences of restoring incremental dumps on top of
full dump work, and to ensure that a subsequent vos release can do a real
incremental release for the restored incremental dumps.

It also fixes the fact that the vos exa command hides potential problems
with timestamps by refusing to display a last updated timestamp that is
older than the creation timestamp.