misc-windows-20040731
Move most log entries to the osi trace log from the afsd_init.log

Add log entries to help diagnose the lost callback problem

Fix a variety of type errors in header files and casts

More fixes for cm_Analyze.  Do not allow a retry to be performed
if the timeLeft on the call has expired.  Forcing a retry resets the
timeout period causing the call to never expire.  The side effect is
that the SMB sessions never close.  When the Windows SMB client times
out it will create a new session at a rate of one per 66 seconds.  This
causes the SMB session limit to be exceeded and we panic.

Modify the SMB Session limit to be four less then the number of SMB
NCBs.  This is because the server itself uses 4 NCBs which are not associated
with sessions.