winxp-sp2-pioctl-20040805
On Windows XP SP2, when a Kerberos principal is used for login and the
KDC is not reachable, Windows will return a DOWNGRADE_DETECTED error
when attempting to open the ioctl file.  This is because NTLM will not
be permitted when there is the potential of an attack.  There is a hack
to get around this.  Manually create a CIFS connection to the AFS client
service while specifying the current user's name.  This will use the cached
password and allow the downgrade to NTLM.