xp-sp2-fix-plus-20040810
The XP SP2 problem was caused by attempting to call bind() from
within DllMain which is no longer permitted.  Added a new function
AfsLogonInit() which is called from every entry point to ensure
that the necessary initializaton is performed.

Cleaned up the prior fix to smb3.c to better test the invalid
conditions.