time_t-pointer-conversions-20040908
FIXES 15169

Modify all parameters to functions which take a pointer to time_t
to actually provide a pointer to time_t instead of pointers to long,
afs_int32, afs_uint32, etc.

This is to provide compatibility with systems that use 64-bit time_t
typedefs such as the IBM compilers for AIX and the Microsoft Visual
Studio 2005 compiler.

This patch does not address any of the issues of data loss caused
by converting 64-bit time_t to 32-bit signed or unsigned values.