cifs-pattern-match-20040921
The pattern matching algorithm was failing to match strings when the
pattern terminated in a '*'.  The logic was also too complex because
it failed to simply the patterns prior to processing.  Any combination
of '*' and '?' == '*' according to the Windows file name pattern
matching rules.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================
FIXES 15365

The pattern matching algorithm was failing to match strings when the
pattern terminated in a '*'.  The logic was also too complex because
it failed to simply the patterns prior to processing.  Any combination
of '*' and '?' == '*' according to the Windows file name pattern
matching rules.