viced-callback-lock-host-during-multi-breakcallback-20041011
since it looks like CheckHost is using a gc'd client conn, this more or less has to be it.

====================
This delta was composed from multiple commits as part of the CVS->Git migration.
The checkin message with each commit was inconsistent.
The following are the additional commit messages.
====================

since it looks like CheckHost is using a gc'd client conn, this more or less has
 to be it.

and don't deadlock.

====================

since it looks like CheckHost is using a gc'd client conn, this more or less has
 to be it.

and don't deadlock.

====================

since it looks like CheckHost is using a gc'd client conn, this more or less has
 to be it.

and don't deadlock.